requirements.txt
fastapi>=0.116.1
langchain>=0.3.27
langchain-community>=0.3.29
langchain-openai>=0.3.33
openai>=1.107.2
pgvector>=0.4.1
psycopg2-binary>=2.9.10
pypdf>=6.0.0
python-multipart>=0.0.20
requests>=2.32.5
sqlalchemy>=2.0.43
streamlit>=1.49.1
uvicorn>=0.35.0
